% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godwit_tracks_gambia.R
\docType{data}
\name{godwit_tracks_gambia}
\alias{godwit_tracks_gambia}
\title{Example godwit tracking dataset from Gambia}
\format{
A data frame with columns:
\describe{
  \item{id}{Unique identifier for each tracked animal}
  \item{ring_id}{Ring identifier}
  \item{sensor}{Sensor type or device identifier}
  \item{sex}{Sex of the individual}
  \item{timestamp}{Datetime of the location fix}
  \item{x}{Longitude coordinate}
  \item{y}{Latitude coordinate}
}
}
\usage{
data(godwit_tracks_gambia)
}
\description{
A dataset of godwit tracking data from Gambia used as an example for ezTrack functions.
}
\keyword{datasets}
