% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godwit_tracks_gambia.R
\docType{data}
\name{godwit_tracks_gambia}
\alias{godwit_tracks_gambia}
\title{Example Godwit Tracking Dataset from Gambia}
\format{
A data frame with columns:
\describe{
  \item{individual.local.identifier}{Unique identifier for each tracked animal}
  \item{timestamp}{Datetime of the location fix}
  \item{location.long}{Longitude coordinate (WGS84)}
  \item{location.lat}{Latitude coordinate (WGS84)}
}
}
\source{
Movebank
}
\usage{
data(godwit_tracks_gambia)
}
\description{
A dataset of godwit tracking data from Gambia used as an example for ezTrack functions.
}
\examples{
data(godwit_tracks_gambia)
godwit_tracks_gambia <- ez_track(godwit_tracks_gambia)
ez_summary(godwit_tracks_gambia)
ez_home_range(godwit_tracks_gambia)
}
\keyword{datasets}
